# apps/api/tests/baseline/openapi_after_step1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudSite API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/resources": {
      "get": {
        "summary": "Resources",
        "operationId": "resources_api_resources_get",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type"
            }
          },
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content Type"
            }
          },
          {
            "name": "folder_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder Id"
            }
          },
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 24,
              "title": "Page Size"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "modified_at",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourcePageOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources/{resource_id}": {
      "get": {
        "summary": "Resource Detail",
        "operationId": "resource_detail_api_resources__resource_id__get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceDetailOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources/{resource_id}/preview": {
      "get": {
        "summary": "Resource Preview Capability",
        "operationId": "resource_preview_capability_api_resources__resource_id__preview_get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources/{resource_id}/text-preview": {
      "get": {
        "summary": "Resource Text Preview",
        "operationId": "resource_text_preview_api_resources__resource_id__text_preview_get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextPreviewOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources/{resource_id}/pdf-preview": {
      "get": {
        "summary": "Resource Pdf Preview",
        "operationId": "resource_pdf_preview_api_resources__resource_id__pdf_preview_get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/resources/{resource_id}/office-preview": {
      "get": {
        "summary": "Resource Office Preview",
        "operationId": "resource_office_preview_api_resources__resource_id__office_preview_get",
        "parameters": [
          {
            "name": "resource_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Resource Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/folders": {
      "get": {
        "summary": "Folders",
        "operationId": "folders_api_folders_get",
        "parameters": [
          {
            "name": "content_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Content Type"
            }
          },
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderListOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/folders/{folder_id}": {
      "get": {
        "summary": "Folder Detail",
        "operationId": "folder_detail_api_folders__folder_id__get",
        "parameters": [
          {
            "name": "folder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Folder Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 24,
              "title": "Page Size"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "name",
              "title": "Sort"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDetailOutput"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BreadcrumbItem": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "BreadcrumbItem"
      },
...
      "FolderDetailOutput": {
        "properties": {
          "folder": {
            "$ref": "#/components/schemas/FolderSummary"
          },
          "breadcrumbs": {
            "items": {
              "$ref": "#/components/schemas/BreadcrumbItem"
            },
            "type": "array",
            "title": "Breadcrumbs"
          },
          "child_folders": {
            "items": {
              "$ref": "#/components/schemas/FolderSummary"
            },
            "type": "array",
            "title": "Child Folders"
          },
          "resources": {
            "$ref": "#/components/schemas/ResourcePageOutput"
          }
        },
        "type": "object",
        "required": [
          "folder",
          "breadcrumbs",
          "child_folders",
          "resources"
        ],
        "title": "FolderDetailOutput"
      },
      "FolderListOutput": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FolderSummary"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "FolderListOutput"
      },
      "FolderSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "depth": {
            "type": "integer",
            "title": "Depth"
          },
          "child_folder_count": {
            "type": "integer",
            "title": "Child Folder Count"
          },
          "resource_count": {
            "type": "integer",
            "title": "Resource Count"
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "parent_id",
          "content_type",
          "depth",
          "child_folder_count",
          "resource_count",
          "modified_at"
        ],
        "title": "FolderSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ParentSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ParentSummary"
      },
...
      "PreviewCapability": {
        "properties": {
          "preview_type": {
            "type": "string",
            "title": "Preview Type"
          },
          "can_preview": {
            "type": "boolean",
            "title": "Can Preview"
          },
          "preview_mode": {
            "type": "string",
            "title": "Preview Mode"
          },
          "browser_native": {
            "type": "boolean",
            "title": "Browser Native",
            "default": false
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type",
            "default": ""
          },
          "extension": {
            "type": "string",
            "title": "Extension",
            "default": ""
          },
          "reason": {
            "type": "string",
            "title": "Reason",
            "default": ""
          },
          "gateway_url": {
            "type": "string",
            "title": "Gateway Url",
            "default": ""
          },
          "can_download": {
            "type": "boolean",
            "title": "Can Download"
          }
        },
        "type": "object",
        "required": [
          "preview_type",
          "can_preview",
          "preview_mode",
          "can_download"
        ],
        "title": "PreviewCapability"
      },
      "ResourceDetailOutput": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParentSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "extension": {
            "type": "string",
            "title": "Extension"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          },
          "thumbnail": {
            "type": "string",
            "title": "Thumbnail",
            "default": ""
          },
          "breadcrumbs": {
            "items": {
              "$ref": "#/components/schemas/BreadcrumbItem"
            },
            "type": "array",
            "title": "Breadcrumbs"
          },
          "related": {
            "items": {
              "$ref": "#/components/schemas/ResourceSummary"
            },
            "type": "array",
            "title": "Related"
          },
          "capabilities": {
            "$ref": "#/components/schemas/PreviewCapability"
          },
          "previous": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "next": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResourceSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "parent_id",
          "content_type",
          "extension",
          "mime_type",
          "size",
          "modified_at",
          "breadcrumbs",
          "related",
          "capabilities"
        ],
        "title": "ResourceDetailOutput"
      },
      "ResourcePageOutput": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResourceSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "title": "ResourcePageOutput"
      },
      "ResourceSummary": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "parent": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ParentSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "extension": {
            "type": "string",
            "title": "Extension"
          },
          "mime_type": {
            "type": "string",
            "title": "Mime Type"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "modified_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Modified At"
          },
          "thumbnail": {
            "type": "string",
            "title": "Thumbnail",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "parent_id",
          "content_type",
          "extension",
          "mime_type",
          "size",
          "modified_at"
        ],
        "title": "ResourceSummary"
      },
...
      "TextPreviewOutput": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content"
          },
          "truncated": {
            "type": "boolean",
            "title": "Truncated"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "encoding": {
            "type": "string",
            "title": "Encoding"
          },
          "preview_type": {
            "type": "string",
            "title": "Preview Type"
          }
        },
        "type": "object",
        "required": [
          "content",
          "truncated",
          "size",
          "encoding",
          "preview_type"
        ],
        "title": "TextPreviewOutput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}